Improved File system and Custom Text
