+++ App.js
@@ -5,8 +5,8 @@ import { persistStore } from 'redux-persist';
 import { PersistGate } from 'redux-persist/integration/react';
 
 import Store from './src/store/index';
-import { RobotoRegular, RobotoBold } from './src/utils/components/AppText';
-import { PRI_COLOR } from './src/utils/constants';
+import { PRI_COLOR } from './src/utils/constants/colors';
+import AppText from './src/utils/components/AppText';
 
 const persistor = persistStore(Store);
 
@@ -24,12 +24,12 @@ export default class App extends Component {
       <Provider store={Store}>
         <PersistGate persistor={persistor} loading={this.renderLoading()}>
           <View style={{justifyContent: 'center', flex: 1, alignItems: 'center', padding: 30}}>
-            <RobotoBold style={{alignSelf: 'center', textAlign: 'center', color: PRI_COLOR}}>
+            <AppText style={{alignSelf: 'center', textAlign: 'center', color: PRI_COLOR }} family='Roboto' variant='Bold'>
               Welcome to E-Practice
-            </RobotoBold>
-            <RobotoRegular style={{alignSelf: 'center', textAlign: 'center'}}>
+            </AppText>
+            <AppText style={{alignSelf: 'center', textAlign: 'center'}} family='Roboto' variant='Regular' >
               We are here to learn while building something spectacular. #EnjoyCoding
-            </RobotoRegular>
+            </AppText>
           </View>
         </PersistGate>
       </Provider>
