+++ package.json
@@ -8,6 +8,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "prop-types": "^15.7.2",
     "react": "16.8.6",
     "react-native": "0.60.0",
     "react-redux": "^6.0.1",
