+++ src/store/index.js
@@ -1,5 +1,5 @@
 import { combineReducers, createStore } from "redux";
-import userReducers from "./reducers/userReducers";
+import userReducers from "./reducers/user.reducer";
 import { persistReducer } from 'redux-persist'
 import storage from 'redux-persist/lib/storage' // defaults to localStorage for web and AsyncStorage for react-native
