--- src/store/initialState.js
@@ -1,7 +1,0 @@
-const InitialState = {
-  user: {},
-  isAuthenticated: false,
-  token: null
-};
-
-export default InitialState;
