--- src/utils/constants.js
@@ -1,3 +1,0 @@
-export const LOGIN = 'LOGIN';
-export const LOGOUT = 'LOGOUT';
-export const PRI_COLOR = '#17a4d6';
