+++ src/utils/constants/redux.js
@@ -1,0 +1,2 @@
+export const LOGIN = 'LOGIN';
+export const LOGOUT = 'LOGOUT';
